backend/requirements.txt
asgiref==3.11.0
boto3==1.42.8
botocore==1.42.8
charset-normalizer==3.4.4
Django==6.0
django-cors-headers==4.9.0
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
jmespath==1.0.1
mysqlclient==2.2.7
pillow==12.0.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
reportlab==4.4.6
s3transfer==0.16.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.4
urllib3==2.6.2
wheel==0.45.1
